+++ xstate/CHANGELOG.md
@@ -7,6 +7,14 @@ The format is based on [Common Changelog](https://common-changelog.org/).
 [1.0.0]: https://github.com/PaulRBerg/prb-effect/releases/tag/%40prb%2Feffect-xstate%401.0.0
 [1.0.1]: https://github.com/PaulRBerg/prb-effect/releases/tag/%40prb%2Feffect-xstate%401.0.1
 [2.0.0]: https://github.com/PaulRBerg/prb-effect/releases/tag/%40prb%2Feffect-xstate%402.0.0
+[2.0.1]: https://github.com/PaulRBerg/prb-effect/releases/tag/%40prb%2Feffect-xstate%402.0.1
+
+## [2.0.1] - 2026-02-17
+
+### Changed
+
+- Prefer confirm-phase `hash` over sign-phase `hash` in `createTxMachine`
+- Allow confirm output to clear stale hash with `hash: null`
 
 ## [2.0.0] - 2026-02-10